add: dm event

+++ src/events/filter.js
@@ -9,6 +9,8 @@ export const name = 'messageCreate'
  * @param {Message} message
  */
 export async function run(message) {
+    if (message.channel.type !== 'text') return
+
     if (filter.isProfane(message.content)) {
         message.delete()
     }
